Override expandBuff for JavaCharStream
The implementation of expandBuff() in AbstractCharStream sets
maxNextCharInd to be the the buffer position. However this buffer
size is not the same as the one of nextCharBuf, which is fixed at
4096. JavaCharStream uses maxCharInd to bound nextCharInd, which
it uses to access nextCharBuf in readByte(). Therefore this can
cause OOB access when JavaCharStream tries to read something big
enough to cause expandBuff() to be called.

This fix simply backports the version of the method that was
present in javacc in JavaCharStream and has it override the
superclass's implementation.